Move components to the components folder
