+++ src/lib/components/withoutLoan/snippets/GetLoanIntentSnippet.svelte
